refactor(backend): split atomic writer integration tests
